refactor (ChoreList) made new component wrapping input fields and submit button
